type pic changed again

+++ Sport/app/src/main/java/com/example/myapplication/adapter/RecordAdapter.java
@@ -60,7 +60,7 @@ public class RecordAdapter extends RecyclerView.Adapter<RecordAdapter.ViewHolder
         ImageView rank_img;
         TextView rank_text;
 
-        int record_id;
+        long record_id;
 
         public ViewHolder(View view) {
             super(view);
